redesign changes
