+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/proteins-list/proteins-list-item/ProteinsListItem.tsx
@@ -21,9 +21,6 @@ import ProteinsListItemInfo from '../proteins-list-item-info/ProteinsListItemInf
 
 import { Transcript } from 'src/content/app/entity-viewer/types/transcript';
 
-import { getCommaSeparatedNumber } from 'src/shared/helpers/formatters/numberFormatter';
-import { getFeatureCoordinates } from 'src/content/app/entity-viewer/shared/helpers/entity-helpers';
-
 import transcriptsListStyles from 'src/content/app/entity-viewer/gene-view/components/default-transcripts-list/DefaultTranscriptsList.scss';
 import styles from './ProteinsListItem.scss';
 
@@ -33,31 +30,19 @@ type Props = {
 };
 
 const ProteinsListItem = (props: Props) => {
-  const { transcript } = props;
-
   const [shouldShowInfo, setShouldShowInfo] = useState(false);
   const toggleListItemInfo = () => setShouldShowInfo(!shouldShowInfo);
-  const getSplicedRNALength = () => {
-    const rnaLength = transcript.exons.reduce((length, exon) => {
-      const { start, end } = getFeatureCoordinates(exon);
-      return length + (end - start + 1);
-    }, 0);
-
-    return getCommaSeparatedNumber(rnaLength);
-  };
 
   const midStyles = classNames(transcriptsListStyles.middle, styles.middle);
 
   return (
     <>
       <div className={transcriptsListStyles.row}>
-        <div className={transcriptsListStyles.left}>UniProt P51587</div>
+        <div className={transcriptsListStyles.left}></div>
         <div onClick={toggleListItemInfo} className={midStyles}>
-          <div>Protein description from UniProt</div>
           <div>{props.transcript.cds?.protein_length} aa</div>
-          <div className={styles.splicedLength}>
-            Spliced RNA length <strong>{getSplicedRNALength()}</strong> bp
-          </div>
+          <div>Protein description from UniProt</div>
+          <div>{props.transcript.product?.stable_id}</div>
         </div>
         <div
           className={transcriptsListStyles.right}
